Bug name changes
Wi-card > Wicard
It could both come from Wi-Fi Card, as well as Wildcard. I removed the - so it could refer to both, and it also looks better.

Serial > Seeria
Its name is a corruption of Serial, so I went with Seeria.

Bandoline > Bandlin
Bandolier > Bandlier
Removed the "o", so it's more in-line with Band. Also removed the "e" from Bandlin's name, otherwise it could be read as "line".

Einisu/Anis > Aynis
Decide to go with "Ay" to make it sound like "エイ".

Opti > Optie
It has a long "i" in its name, so Optie is better.

SDcard > S-Dcard
Moved the -, so that it looks more in-line with the other evolutions.

Purein/Pyuureen > Purane
Likely comes from Pure and Lane, went with Purane both to make it more similar to Pre, as well as reference "lane" directly.

Debackin > Debuckin
Changed from "a" to "u", so it would be more similar to Debug.

Hnas > HNAS
Went all caps since it's likely an acronym.

Hnonius > HNonius
Capitalization to make it look more similar to HNAS

Recover > Recove
It's a corruption of recover, not the word itself.

Aspl > Aspla
Lacking the "a" on its end changes how it's pronounced.

Toranza > Transa
Toranje > Transe
Removed the "o", so it's more in-line with Tran. The ending syllables were replaced to make it sound better in english.

Marubo > Malbo
It likely comes from Malware, so "maru" was replaced by "mal"

Raytrace > Raytray
It's a corruption of raytrace, not the word itself.

Tabley > Tablay
Changed from "e" to "a" so it could preserve the japanese sound better.

Cook Gen > Cook Generay
It comes from Generate, but "Gen" lacks part of the original name (Kukku Jenere). Generay preserves that.